backend/requirements.txt
# FastAPI and web server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# Data validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# CORS
python-dotenv==1.0.1

# Gemini AI
google-genai>=1.0.0
